WIP: Major refactor of map movement logic. There's a pretty big bug when walking up though. Committing this now so I can get back to a 75% working state at least.

+++ src/game/objects/image.cpp
@@ -7,12 +7,12 @@ namespace Game {
             this->position = {0, 0, 0, 0};
         }
 
-        Image::Image(std::shared_ptr<Resources::Texture> texture, const int x, const int y) : Image() {
+        Image::Image(std::shared_ptr<Resources::Texture> texture, const float x, const float y) : Image() {
             this->SetTexture(texture);
             this->SetPosition(x, y);
         }
 
-        Image::Image(const std::string& textureName, const int x, const int y) : Image() {
+        Image::Image(const std::string& textureName, const float x, const float y) : Image() {
             this->SetTexture(Services::Locator::TextureService()->GetTexture(textureName));
             this->SetPosition(x, y);
         }
